Moved json types declaration to the separate mod
